refactor(wkt): split float serializers
The `internal.rs` file was getting too long.
